packages/api/requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
azure-identity==1.17.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.30.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
